--- StyleHelper.js
@@ -1,21 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-export const colors = {
-    screenBackground: '#AAA8C8',
-    darkpurple: '#3B387E',
-};
-export default GlobalStyles = StyleSheet.create({
-       ScreenContainer: {
-        flex: 1,
-        backgroundColor: colors.screenBackground,
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    lable: {
-        fontSize: 15,
-        color: colors.darkpurple,
-        fontWeight: 'bold',
-    },
-    
-
-})
